pyproject.toml
[tool.poetry]
name = "bewise-test-task-2"
version = "0.1.0"
description = "test task from bewise"
authors = ["KLYMENKORUS <klimenkoruslan11@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.19"
asyncpg = "^0.28.0"
fastapi-users = "^12.1.0"
fastapi-users-db-sqlalchemy = "^6.0.0"
uvicorn = "^0.23.1"
pydub = "^0.25.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.2"
jinja2 = "^3.1.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"
aiofiles = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
